Build with --mixed on i386 to be compatible with amd64.
jackd2 on amd64 is built with --mixed turned on, so a 32bit library will
be created during the amd64 build process. However, we don't ship this
i386 library in the amd64 package but expect users to install the i386
version from the archive (apt-ia32 or chroot approach).

The i386 lib uses a different memory layout (4-byte align instead of
8-byte), thus breaking the whole magic. With --mixed also enabled on
i386, the i386 lib will use the same alignment as on amd64. Though this
might slightly increase the runtime memory footprint, it at least
enables users to run i386 jack clients on amd64 jackd servers.

We might want to have a look at this issue again once the new multiarch
strategy is fully set.