Override lintian error about libjacknet
libjacknet is currently a jackd2 only library that refactors part of
netjack1 into a library.

Theoretically, this could be useful for jackd1 too, but this requires
massive upstream involvement.

We hence ship libjacknet in libjack-jackd2-0 and pretend that it's ok
not to mention it in the corresponding shlibs file.

The clean approach would probably be to introduce a new libjacknet
package, but since there is only one user, I feel it's a bit too early
for this move.