Don't fail if waflib cannot be removed
This is probably only relevant for people who use debcheckout to build
the package. In this case, patches are not yet applied when

   debian/rules clean

is called. Since waflib/ is missing (it's applied from waf.patch), find
fails and terminates the build process.

This patch ignores any errors from this single find command.